Move Cube robot to utils test package
